+++ Editor/CASEditorSettings.cs
@@ -22,7 +22,7 @@ namespace CAS.UEditor
         public string mostPopularCountryOfUsers = "BR";
 
         public bool multiDexEnabled = true;
-        public bool permissionAdIdRemoved = false;
+        public Permission permissionAdId = Permission.Auto;
         [Obsolete("Exo player used in any case")]
         public bool exoPlayerIncluded = true;
         public bool generateAndroidQuerriesForPromo = true;
@@ -41,13 +41,28 @@ namespace CAS.UEditor
         public KeyValuePair[] userTrackingUsageDescription = new KeyValuePair[0];
 
 
-        public static CASEditorSettings Load( bool createAsset = false )
+        public static CASEditorSettings Load(bool createAsset = false)
         {
-            var asset = CASEditorUtils.GetSettingsAsset( "CASEditorSettings",
-                CASEditorUtils.editorFolderPath, typeof( CASEditorSettings ), createAsset, null );
+            var asset = CASEditorUtils.GetSettingsAsset("CASEditorSettings",
+                CASEditorUtils.editorFolderPath, typeof(CASEditorSettings), createAsset, null);
             if (asset)
-                return ( CASEditorSettings )asset;
+                return (CASEditorSettings)asset;
             return CreateInstance<CASEditorSettings>();
+        }
+
+        public bool isUseAdvertiserIdLimited(Audience audience)
+        {
+            var permission = permissionAdId;
+            if (permission == Permission.Auto)
+                return audience == Audience.Children;
+            return permission == Permission.Removed;
+        }
+
+        public enum Permission
+        {
+            Auto = 0,
+            Required = 1,
+            Removed = 2
         }
     }
 }